Create PausableTimerListener abstract class
This class implements TimerListener and adds methods that allow the Timer
to start() and stop() by default. It also instantiates the Timer inside the
abstract class, so it does not need to be called in every separate constructor.
The possible downside of this is that Timer refresh rates need to be uniformized.
Thus, each separate Timer-based class in our final project should extend
PausableTimerListener instead of implementing TimerListener.